+++ src/common/i18n/en.js
@@ -41,7 +41,8 @@ Validity.i18n = {
                 text_empty: 'There is currently not any available validity period.<br />'
                     +'You can try to define a new one by decreasing some registered validity period(s).',
                 text_one: 'The following periods are not covered by any validity record.<br />'
-                    +'If this is not what you want, you have to increase some validity periods.',
+                    +'This is not forbidden in any case, and this also may be what you actually want, and consistent with your use case.<br />'
+                    +'If this is not what you want, you may want to increase the validity of some periods, or define new validity periods.',
                 title_mergeleft: 'Merging with previous period',
                 title_mergeright: 'Merging with next period',
                 title_remove: 'Removing a period',